Introduce IAutoRestCodeGeneratorFactory
