+++ src/ApiClientCodeGen.CLI/Commands/AutoRestCommand.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core.Generators;
-using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core.Generators.AutoRest;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Core.Options.AutoRest;
 using McMaster.Extensions.CommandLineUtils;
 
@@ -12,22 +11,22 @@ namespace ApiClientCodeGen.CLI.Commands
     {
         private readonly IAutoRestOptions options;
         private readonly IProcessLauncher processLauncher;
-        private readonly ICodeGeneratorCommandFactory commandFactory;
+        private readonly IAutoRestCodeGeneratorFactory factory;
 
         public AutoRestCommand(
             IConsoleOutput console,
             IAutoRestOptions options,
             IProcessLauncher processLauncher,
             IProgressReporter progressReporter,
-            ICodeGeneratorCommandFactory commandFactory) : base(console, progressReporter)
+            IAutoRestCodeGeneratorFactory factory) : base(console, progressReporter)
         {
             this.options = options ?? throw new ArgumentNullException(nameof(options));
             this.processLauncher = processLauncher ?? throw new ArgumentNullException(nameof(processLauncher));
-            this.commandFactory = commandFactory ?? throw new ArgumentNullException(nameof(commandFactory));
+            this.factory = factory ?? throw new ArgumentNullException(nameof(factory));
         }
 
         public override ICodeGenerator CreateGenerator()
-            => commandFactory.Create<AutoRestCommand>(
+            => factory.Create(
                 SwaggerFile,
                 DefaultNamespace,
                 options,
